move header component to its own folder
