+++ src/pages/GardenView/index.tsx
@@ -1,4 +1,4 @@
-import { Header } from "../../components/Header";
+import { Header } from "../../components/Header/Header";
 
 export const GardenView = ()  => {
   return (
